conformance/bindingTesting/queues/cl_commandQueue_enqueueWriteImage.html

<!--
 Copyright (C) 2013 Samsung Electronics Corporation. All rights reserved.

 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided the following condition
 are met:

 1.  Redistributions of source code must retain the above copyright
     notice, this list of conditions and the following disclaimer.

 2.  Redistributions in binary form must reproduce the above copyright
     notice, this list of conditions and the following disclaimer in the
     documentation and/or other materials provided with the distribution.

 THIS SOFTWARE IS PROVIDED BY SAMSUNG ELECTRONICS CORPORATION AND ITS
 CONTRIBUTORS "AS IS", AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING
 BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
 FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SAMSUNG
 ELECTRONICS CORPORATION OR ITS CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
 INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES(INCLUDING
 BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 DATA, OR PROFITS, OR BUSINESS INTERRUPTION), HOWEVER CAUSED AND ON ANY THEORY
 OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT(INCLUDING
 NEGLIGENCE OR OTHERWISE ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
 EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 -->
<html>
<head>
<meta charset="utf-8">
    <title>WebCLCommandQueue enqueueWriteImage conformance test.</title>
    <link rel="stylesheet" href="../../../resources/js-test-style.css"/>
    <script src="../../../resources/js-test-pre.js"></script>
    <script src="../../../resources/cl_utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>

try {
    debug("WebCLCommandQueue :: enqueueWriteImage() test case. \n");
    debug("( WebCLImage image, CLboolean blockingWrite, sequence<CLuint> origin, sequence<CLuint> region, CLuint hostRowPitch, ArrayBufferView hostPtr, optional sequence<WebCLEvent>? eventWaitList, optional WebCLEvent? event)\n");

    var wtu = WebCLTestUtils;
    var webCLContext = wtu.createContext();
    var webCLCommandQueue = wtu.createCommandQueue(webCLContext);
    var imageWidth = 10;
    var imageHeight = 10;
    var origin = [0, 0];
    var region = [imageWidth, imageHeight];
    var hostRowPitch = 0;
    var hostPtr = new Float32Array(imageWidth * imageHeight);
    for (var i = 0; i < imageWidth * imageHeight; i++) {
        hostPtr[i] = Math.random();
    }

    // MEM_READ_ONLY images.
    var webCLImageDescriptorObject  = wtu.getSupportedImageFormats(webCLContext, webcl.MEM_READ_ONLY, imageWidth, imageHeight);
    var webCLImageRead = wtu.createImage(webCLContext, webcl.MEM_READ_ONLY, webCLImageDescriptorObject);
    shouldBeUndefined("webCLCommandQueue.enqueueWriteImage(webCLImageRead, true, origin, region, hostRowPitch, hostPtr);");
    shouldBeUndefined("webCLCommandQueue.enqueueWriteImage(webCLImageRead, false, origin, region, hostRowPitch, hostPtr);");

    // MEM_WRITE_ONLY images.
    webCLImageDescriptorObject = wtu.getSupportedImageFormats(webCLContext, webcl.MEM_WRITE_ONLY, imageWidth, imageHeight);
    var webCLImageWrite = wtu.createImage(webCLContext, webcl.MEM_WRITE_ONLY, webCLImageDescriptorObject);
    shouldBeUndefined("webCLCommandQueue.enqueueWriteImage(webCLImageWrite, true, origin, region, hostRowPitch, hostPtr);");
    shouldBeUndefined("webCLCommandQueue.enqueueWriteImage(webCLImageWrite, false, origin, region, hostRowPitch, hostPtr);");

    // MEM_READ_WRITE images.
    webCLImageDescriptorObject = wtu.getSupportedImageFormats(webCLContext, webcl.MEM_READ_WRITE, imageWidth, imageHeight);
    var webCLImageReadWrite = wtu.createImage(webCLContext, webcl.MEM_READ_WRITE, webCLImageDescriptorObject);
    shouldBeUndefined("webCLCommandQueue.enqueueWriteImage(webCLImageReadWrite, true, origin, region, hostRowPitch, hostPtr);");
    shouldBeUndefined("webCLCommandQueue.enqueueWriteImage(webCLImageReadWrite, false, origin, region, hostRowPitch, hostPtr);");

    // Testing for Exceptions.
    shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(invalid_object, true, origin, region, hostRowPitch, hostPtr);", "INVALID_MEM_OBJECT");

    var originInvalid = [];
    shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(webCLImageReadWrite, true, originInvalid, region, hostRowPitch, hostPtr);", "INVALID_VALUE");
    originInvalid = [imageWidth + 1, imageHeight + 1];
    shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(webCLImageReadWrite, true, originInvalid, region, hostRowPitch, hostPtr);", "INVALID_VALUE");
    originInvalid = [0];
    shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(webCLImageReadWrite, true, originInvalid, region, hostRowPitch, hostPtr);", "INVALID_VALUE");

    var regionInvalid = [];
    shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(webCLImageReadWrite, true, origin, regionInvalid, hostRowPitch, hostPtr);", "INVALID_VALUE");
    regionInvalid = [imageWidth + 1, imageHeight + 1];
    shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(webCLImageReadWrite, true, origin, regionInvalid, hostRowPitch, hostPtr);", "INVALID_VALUE");
    regionInvalid = [imageWidth]; 
    shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(webCLImageReadWrite, true, origin, regionInvalid, hostRowPitch, hostPtr);", "INVALID_VALUE");

    var hostRowPitchValid = imageWidth * 2;
    shouldBeUndefined("webCLCommandQueue.enqueueWriteImage(webCLImageReadWrite, true, origin, region, hostRowPitchValid, hostPtr);");
    var hostRowPitchInvalid = (Float32Array.BYTES_PER_ELEMENT * imageWidth) - 1;
    shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(webCLImageReadWrite, true, origin, region, hostRowPitchInvalid, hostPtr);", "INVALID_VALUE");
    hostRowPitchInvalid = hostPtr.BYTES_PER_ELEMENT - 1;
    shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(webCLImageReadWrite, true, origin, region, hostRowPitchInvalid, hostPtr);", "INVALID_VALUE");

    var hostPtrInvalid = new Float32Array((imageWidth * imageWidth) - 1);
    shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(webCLImageReadWrite, true, origin, region, hostRowPitch, hostPtrInvalid);", "INVALID_VALUE");

    var webCLContextNew = wtu.createContext();
    var webCLCommandQueueNew = wtu.createCommandQueue(webCLContextNew);
    var webCLImageNew = wtu.createImage(webCLContextNew, webcl.MEM_READ_WRITE, webCLImageDescriptorObject);
    shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(webCLImageNew, true, origin, region, hostRowPitch, hostPtr);", "INVALID_CONTEXT");
    shouldThrowExceptionName("webCLCommandQueueNew.enqueueWriteImage(webCLImageReadWrite, true, origin, region, hostRowPitch, hostPtr);", "INVALID_CONTEXT");

    var webCLEvent = wtu.createEvent(webCLContext);
    var webCLEventNew = wtu.createEvent(webCLContextNew);
    shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(webCLImageReadWrite, true, origin, region, hostRowPitch, hostPtr, [webCLEventNew], null);", "INVALID_CONTEXT");
    shouldThrowExceptionName("webCLCommandQueueNew.enqueueWriteImage(webCLImageReadWrite, true, origin, region, hostRowPitch, hostPtr, [webCLEvent], null);", "INVALID_CONTEXT");
    shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(webCLImageReadWrite, true, origin, region, hostRowPitch, hostPtr, [], webCLEventNew);", "INVALID_CONTEXT");
    shouldThrowExceptionName("webCLCommandQueueNew.enqueueWriteImage(webCLImageReadWrite, true, origin, region, hostRowPitch, hostPtr, [], webCLEvent);", "INVALID_CONTEXT");

    wtu.release(webCLEventNew);
    shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(webCLImageReadWrite, true, origin, region, hostRowPitch, hostPtr, [webCLEventNew], webCLEvent);", "INVALID_EVENT_WAIT_LIST");

    wtu.setStatus(webCLEvent, -1);
    shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(webCLImageReadWrite, true, origin, region, hostRowPitch, hostPtr, [webCLEvent]);", "EXEC_STATUS_ERROR_FOR_EVENTS_IN_WAIT_LIST");

    // FIXME Need to add test case for INVALID_IMAGE_SIZE.

} catch (e) {
    testFailed(e.description);
}
successfullyParsed = true;
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
